fix: fixed tracker issues
