+++ deep_sort/linear_assignment.py
@@ -1,6 +1,7 @@
 # vim: expandtab:ts=4:sw=4
 from __future__ import absolute_import
 import numpy as np
+import copy
 
 from scipy.optimize import linear_sum_assignment  # <-- CHANGE HERE
 
@@ -58,14 +59,14 @@ def min_cost_matching(
         tracks, detections, track_indices, detection_indices)
     cost_matrix[cost_matrix > max_distance] = max_distance + 1e-5
 
-    cost_matrix_ = cost_matrix.copy()
-    
+    cost_matrix_ = copy.deepcopy(cost_matrix)
+
     try:
         row_indices, col_indices = linear_sum_assignment(cost_matrix_)
     except ValueError:
         print('ERROR: linear_sum_assignment failed', cost_matrix_)
-        cost_matrix_ = distance_metric(tracks, detections, track_indices, detection_indices)
-
+        cost_matrix = distance_metric(
+            tracks, detections, track_indices, detection_indices)
 
     indices = list(zip(row_indices, col_indices))
