+++ scripts/eval.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-BENCHMARK="MOT17" 
+BENCHMARK="MOT20" 
 
 # Base Command
 BASE_CMD="python TrackEval/scripts/run_mot_challenge.py"
